+++ mwes/mwe_05.jl
@@ -5,4 +5,4 @@ segments=se().segments
 state=demo_state_4p(segments+1)
 update_system(viewer, state, kite_scale=0.25)
 sleep(5)
-close(viewer.screen)
+#close(viewer.screen)