+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 - add menu time_lapse for changing the simulation/ replay speed
 - add field `mod_text` to Viewer3D; it defines the quotient of the update rate and the text update rate
+- add textbox `t_sim` to enter the simulation time
 
 ### KiteViewers v0.4.10 - 2024-03-30